+++ front-react/src/components/Card/CardExpTable.jsx
@@ -1,13 +1,32 @@
+/*!
+
+=========================================================
+* Light Bootstrap Dashboard React - v1.3.0
+=========================================================
+
+* Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
+* Copyright 2019 Creative Tim (https://www.creative-tim.com)
+* Licensed under MIT (https://github.com/creativetimofficial/light-bootstrap-dashboard-react/blob/master/LICENSE.md)
+
+* Coded by Creative Tim
+
+=========================================================
+
+* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+*/
 import React, { Component } from "react";
 
-export class Card extends Component {
+export class CardExpTable extends Component {
   render() {
     return (
       <div
         className={"card" + (this.props.plain ? " card-plain" : "")}>
         <div className={"header" + (this.props.hCenter ? " text-center" : "")}>
-          <h4 className="title">{this.props.title}</h4>
-          <p className="category">{this.props.category}</p>
+          <h3 className="title">{this.props.category} {this.props.title}</h3>
+          <br />
+          <p className="category">{this.props.content2}</p>
+          
         </div>
         <div
           className={
@@ -33,4 +52,4 @@ export class Card extends Component {
   }
 }
 
-export default Card;
+export default CardExpTable;