+++ front-react/src/components/ExpenseUpdater/ExpenseUpdater1.jsx
@@ -44,7 +44,6 @@ function ExpenseUpdater1(props) {
 
   function submitExpense() {
     const expenseObj = { amount, name, type };
-    console.log("hello");
     Promise.all([
       axios.put(`http://localhost:8001/api/expenses/add`, expenseObj)
     ])