+++ front-react/src/components/Goal/new.jsx
@@ -1,15 +1,7 @@
-import React, { Component } from "react";
+import React from "react";
 
 export default function New(props) {
-  // function typeDetailer(type) {
-  //   switch (type) {
-  //     case "SFP":
-  //       return "saving for purchase"
-  //     default:
-  //       return type
-  //   }
-  // }
-
+ 
   return (
     <article className="goalNew">
       <div className="icons">