+++ front-react/src/components/MonthPicker/MonthPicker.jsx
@@ -38,8 +38,7 @@ class List extends Component {
         super(props, context)
 
         this.state = {
-            mvalue: {year: 2020, month: 1},
-            // mvalue2: {year: 2020, month: 12},
+            mvalue: props.currentMonth,
             mrange: {from: {year: 2019, month: 8}, to: {year: 2021, month: 12}},
             // mrange2: {from: {year: 2013, month: 11}, to: {year: 2016, month: 3}},
         }
@@ -48,17 +47,10 @@ class List extends Component {
         this.handleAMonthChange = this.handleAMonthChange.bind(this)
         this.handleAMonthDissmis = this.handleAMonthDissmis.bind(this)
 
-        this.handleClickMonthBox2 = this.handleClickMonthBox2.bind(this)
-        this.handleAMonthChange2 = this.handleAMonthChange2.bind(this)
-        this.handleAMonthDissmis2 = this.handleAMonthDissmis2.bind(this)
-
         this._handleClickRangeBox = this._handleClickRangeBox.bind(this)
         this.handleRangeChange = this.handleRangeChange.bind(this)
         this.handleRangeDissmis = this.handleRangeDissmis.bind(this)
 
-        this._handleClickRangeBox2 = this._handleClickRangeBox2.bind(this)
-        this.handleRangeChange2 = this.handleRangeChange2.bind(this)
-        this.handleRangeDissmis2 = this.handleRangeDissmis2.bind(this)
     }
 
     componentWillReceiveProps(nextProps){
@@ -74,9 +66,7 @@ class List extends Component {
             from: 'From', to: 'To',
         }
         const mvalue = this.state.mvalue
-            , mvalue2 = this.state.mvalue2
             , mrange = this.state.mrange
-            , mrange2 = this.state.mrange2
 
         const makeText = m => {
             if (m && m.year && m.month) return (pickerLang.months[m.month-1] + '. ' + m.year)
@@ -86,7 +76,7 @@ class List extends Component {
         return (
             <ul>
                 <li>
-                    <label>Change Month</label>
+                    <h4>Change Month</h4>
                     <div className="edit">
                         <Picker
                             ref="pickAMonth"
@@ -95,8 +85,11 @@ class List extends Component {
                             lang={pickerLang.months}
                             onChange={this.handleAMonthChange}
                             onDismiss={this.handleAMonthDissmis}
+                            chgMonth={this.props.chgMonth}
                         >
-                            <MonthBox value={makeText(mvalue)} onClick={this.handleClickMonthBox} />
+                            <MonthBox value={makeText(mvalue)}
+                             chgMonth={this.props.chgMonth}
+                             onClick={this.handleClickMonthBox} />
                         </Picker>
                     </div>
                 </li>
@@ -109,20 +102,11 @@ class List extends Component {
         this.refs.pickAMonth.show()
     }
     handleAMonthChange(e, value, text) {
-      console.log(value)
+      
     }
     handleAMonthDissmis(value) {
         this.setState( {mvalue: value} )
-    }
-
-    handleClickMonthBox2(e) {
-        this.refs.pickAMonth2.show()
-    }
-    handleAMonthChange2(value, text) {
-        //
-    }
-    handleAMonthDissmis2(value) {
-        this.setState( {mvalue2: value} )
+        this.props.chgMonth(value)
     }
 
     _handleClickRangeBox(e) {
@@ -135,15 +119,6 @@ class List extends Component {
         this.setState( {mrange: value} )
     }
 
-    _handleClickRangeBox2(e) {
-        this.refs.pickRange2.show()
-    }
-    handleRangeChange2(value, text, listIndex) {
-        //
-    }
-    handleRangeDissmis2(value) {
-        this.setState( {mrange2: value} )
-    }
 }
 
 class Main extends Component {
@@ -162,10 +137,9 @@ class Main extends Component {
     }
 
     render() {
-
         return (
             <div className="list-area">
-                <List />
+                <List {...this.props} />
             </div>
         )
     }
@@ -173,8 +147,6 @@ class Main extends Component {
 
 
 
-export default () => {
-  
-  return (<Main/>);
-
+export default (props) => {
+  return (<Main {...props} />);
 }