+++ front-react/src/hooks/reducers/app.js
@@ -1,11 +1,11 @@
 const SET_DATA = "SET_DATA";
-
-export { SET_DATA };
+const SET_DATE = "SET_DATE";
+const SET_EDU_ANSWERS = "SET_EDUCATION_ANSWERS"
+const SET_EDU_PROGRESS = "SET_EDU_PROGRESS"
 
 export default function reducerz(state, action) {
   switch (action.type) {
     case SET_DATA:
-      console.log("running SET_DATA: ", action);
       return {
         ...state,
         expenses: action.expenses,
@@ -14,9 +14,34 @@ export default function reducerz(state, action) {
         goals: action.goals,
         users: action.users
       };
+    case SET_DATE:
+      console.log(action, 'actionzz')
+      return {
+        ...state,
+        date: action.date
+      }; 
+    case SET_EDU_ANSWERS:
+    console.log(action, 'edu actionzz')
+    return {
+      ...state,
+      educationAnswers: action.educationAnswers
+    };  
+    case SET_EDU_PROGRESS:
+    console.log(action, 'edu actionzz')
+    return {
+      ...state,
+      eduProgress: action.eduProgress
+    };   
     default:
       throw new Error(
         `Tried to reduce with unsupported action type: ${action.type}`
       );
   }
 }
+
+export {
+  SET_DATA,
+  SET_DATE,
+  SET_EDU_ANSWERS,
+  SET_EDU_PROGRESS
+};