+++ front-react/src/variables/EducationArticles.jsx
@@ -1,0 +1,59 @@
+
+
+const articles=[{
+  id:1,
+  title:"Learn about MONEY",
+  link:"https://www.moneyadviceservice.org.uk/en/articles/beginners-guide-to-managing-your-money",
+  image:"https://i.ytimg.com/vi/4N1ayk-Xba4/maxresdefault.jpg",
+  question:"What is money?",
+  answer:1,
+  a1:"A Medium of exchange, allows people to exchange goods & services",
+  a2:"Dollar Dollar Bills!",
+  a3:"Gold coins",
+  a4:"Ones, and Zeros on a server somewhere"
+  },
+  {
+    id:2,
+    title:"TFSA VS RRSP",
+    link:"https://help.wealthsimple.com/hc/en-ca/articles/115015719107-Should-I-use-a-TFSA-or-an-RRSP-",
+    image:"https://i.ytimg.com/vi/4N1ayk-Xba4/maxresdefault.jpg",
+    question:"What does TFSA stand for?",
+    answer:2,
+    a1:"Tax Fund Savings Account",
+    a2:"Tax Free Savings Account",
+    a3:"Triple Fund Super Admin",
+    a4:"Tractor FUN Super AnT"
+  },
+  {
+    id:3,
+    title:"Stocks/Bonds/Mutual Funds?!?",
+    link:"https://blog.vancity.com/intro-investing-stocks-bonds-mutual-funds/",
+    image:"https://i.ytimg.com/vi/4N1ayk-Xba4/maxresdefault.jpg",
+    question:"What is a bond?",
+    answer:3,
+    a1:"James Bond",
+    a2:"A Loving Realationship",
+    a3:"A bond is a fixed income instrument that represents a loan made by an investor to a borrower",
+    a4:"A bond is a equity instrument that represents a loan made by an investor to a borrower"
+  },
+  {
+    id:4,
+    title:"LEARNORE",
+    link:"https://www.moneyadviceservice.org.uk/en/articles/beginners-guide-to-managing-your-money",
+    image:"https://i.ytimg.com/vi/4N1ayk-Xba4/maxresdefault.jpg"
+  },
+  {
+    id:5,
+    title:"YAH YAH ",
+    link:"https://www.moneyadviceservice.org.uk/en/articles/beginners-guide-to-managing-your-money",
+    image:"https://i.ytimg.com/vi/4N1ayk-Xba4/maxresdefault.jpg"
+  }
+
+
+
+]
+
+
+module.exports = {
+  articles
+};